add check if PID var is empty

+++ bash_logout
@@ -8,5 +8,7 @@ fi
 
 # kill remaining ssh-agent
 # http://mah.everybody.org/docs/ssh
-kill ${SSH_AGENT_PID}
+if [ -n "${SSH_AGENT_PID}" ]; then
+    kill ${SSH_AGENT_PID}
+fi
